adding functional tests that use CasperJs to interact with the UI (view.js)
