+++ package.json
@@ -29,6 +29,7 @@
   "homepage": "https://github.com/bicarbon8/todoTxtWebUi",
   "devDependencies": {
     "grunt": "~0.4.5",
+    "grunt-casperjs": "^2.1.0",
     "grunt-contrib-clean": "~0.6.0",
     "grunt-contrib-cssmin": "^0.11.0",
     "grunt-contrib-jshint": "~0.10.0",
