--- tests/allTests.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<!--
-/**********************************************************************
- * This javascript was created according to the specifications at
- * http://todotxt.com/ and is intended to allow users to access their
- * todo.txt files in a user-friendly and easy to visualize manner.
- *
- * Once initially uploaded, the todo.txt file will
- * be loaded into an HTML5 localStorage and managed from there.
- * The web page then allows downloading changes back to the user
- * in a txt format compliant with the todo.txt specifications, but
- * having re-sorted the tasks.
- * 
- * @Created: 08/14/2012
- * @Author: Jason Holt Smith (bicarbon8@gmail.com)
- * @Version: 0.0.1
- * Copyright (c) 2012 Jason Holt Smith. todoTxtWebUi is distributed under
- * the terms of the GNU General Public License.
- * 
- * This file is part of todoTxtWebUi.
- * 
- * todoTxtWebUi is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * todoTxtWebUi is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with todoTxtWebUi.  If not, see <http://www.gnu.org/licenses/>.
- **********************************************************************/
--->
-<html>
-<head>
-<meta charset="utf-8">
-<title>TodoTxtWebUi Tests</title>
-<link rel="stylesheet" href="http://code.jquery.com/qunit/qunit-1.15.0.css">
-<script language="javascript" src="http://code.jquery.com/jquery-1.9.1.min.js"></script>
-</head>
-<body>
-<div id="qunit"></div>
-<div id="qunit-fixture"></div>
-<script src="http://code.jquery.com/qunit/qunit-1.15.0.js"></script>
-<script src="deps/qunit-parameterize-0.4.js"></script>
-<script src="../js/todoTxt.js"></script>
-<script src="../js/objects/utils.js"></script>
-<script src="../js/classes/task.js"></script>
-<script src="js/todoTxtTests.js"></script>
-<script src="js/todoTxtTaskTests.js"></script>
-</body>
-</html>
